Fix local dev env settings
